requirements.txt
uvicorn
pydantic
python-jose
jwt
python-dotenv
python-multipart
pyjwt
python-jwt